fix: Properly segregate the new editor window classes into their own files and add documentation to them as-is appropriate.
