+++ Assets/Plugins/Source/Editor/EditorWindows/EncryptionKeyWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 301e1682c1a1e1b4eb82e35e0f6a7d30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
